requirements.txt
appdirs==1.4.3
asgiref==3.2.7
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
cloudinary==1.21.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.14
django-allauth==0.41.0
django-cloudinary-storage==0.2.3
django-crispy-forms==1.9.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
gunicorn==19.6.0
idna==2.9
oauthlib==3.1.0
packaging==20.3
Pillow==7.0.0
psycopg2==2.8.5
pyparsing==2.4.6
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
uWSGI==2.0.19.1
whitenoise==5.0.1
